+++ src/main/java/com/utc2/petShop/controllers/controllerPet.java
@@ -1,20 +1,30 @@
-package petShop.controllers;
+package com.utc2.petShop.controllers;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.*;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.StackPane;
+import javafx.scene.layout.VBox;
 
 import java.net.URL;
 import java.util.ResourceBundle;
 
-public class controllerHome implements Initializable {
+public class controllerPet implements Initializable {
 
     @FXML
-    private TreeView<String> tree;
+    TreeView<String> tree;
 
     @FXML
     private ScrollPane SrcollLeft,ScrollCenter;
+
+    @FXML
+    private HBox HBoxSrcollCenter;
+
+    @FXML
+    private VBox VBoxSrcollCenter;
 
     @FXML
     private Button buttonFind;
@@ -33,6 +43,40 @@ public class controllerHome implements Initializable {
 
     @FXML
     private Button buttonForward;
+
+    @FXML
+    private Label labelAge;
+
+    @FXML
+    private Label labelBreed;
+
+    @FXML
+    private Label labelFurColor;
+
+    @FXML
+    private Label labelGender;
+
+    @FXML
+    private Label labelGiaBan;
+
+    @FXML
+    private Label labelHealthStatus;
+
+    @FXML
+    private Label labelMoTa;
+
+    @FXML
+    private Label labelName;
+
+    @FXML
+    private Label labelPetID;
+
+    @FXML
+    private Label labelVaccinated;
+
+    @FXML
+    private Label labelWeight;
+
 
     public void actionFind(ActionEvent event) {
         System.out.println(textSreach.getText());
@@ -54,6 +98,7 @@ public class controllerHome implements Initializable {
     public void actionForward(ActionEvent event) {
         //sử lý sự kiện Forward
     }
+
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
@@ -78,5 +123,16 @@ public class controllerHome implements Initializable {
         tree.getSelectionModel().selectedItemProperty().addListener((observable, oldValue, newValue) -> {
         // rằng buộc cho TreeView
         });
+        labelGiaBan.setText("đưa giá vào đây");
+        labelMoTa.setText("đưa mô tả vào đây");
+        labelName.setText("đưa tên vào đây");
+        labelPetID.setText("đưa Pet ID vào đây");
+        labelAge.setText("đưa tuổi vào đây");
+        labelGender.setText("đưa giới tính vào đây");
+        labelVaccinated.setText("đưa Vaccinated vào đây");
+        labelWeight.setText("đưa cân nặng vào đây");
+        labelHealthStatus.setText("đưa tình trạng sức khoẻ vào đây");
+        labelFurColor.setText("đưa màu lông vào đây");
+        labelBreed.setText("đưa gống vào đây");
     }
 }
