+++ src/main/java/com/utc2/petShop/dataBase/dataBaseConnection.java
@@ -1,0 +1,4 @@
+package com.utc2.petShop.dataBase;
+
+public class dataBaseConnection {
+}
