+++ src/main/java/com/utc2/petShop/model/Bill.java
@@ -1,4 +1,4 @@
-package petShop.model;
+package com.utc2.petShop.model;
 
 import java.util.Date;
