+++ src/main/java/com/utc2/petShop/model/Pet.java
@@ -1,4 +1,4 @@
-package petShop.model;
+package com.utc2.petShop.model;
 
 // Represents a pet in the store
 public class Pet {
