+++ src/main/java/com/utc2/petShop/model/Product.java
@@ -1,4 +1,4 @@
-package petShop.model;
+package com.utc2.petShop.model;
 
 // Represents products sold in the pet store
 public class Product {
