+++ src/main/java/com/utc2/petShop/model/User.java
@@ -1,4 +1,4 @@
-package petShop.model;
+package com.utc2.petShop.model;
 
 // Represents a user in the system (customer, employee, manager)
 public class User {
