+++ src/main/resources/views/sampleProgressBar.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.shape.Line?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="510.0" prefWidth="1029.0" style="-fx-background-color: dcdcdc;" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="petShop.controllers.controllerProgressBar">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="510.0" prefWidth="1029.0" style="-fx-background-color: dcdcdc;" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.utc2.petShop.controllers.controllerProgressBar">
    <children>
       <ProgressBar fx:id="probar" layoutX="99.0" layoutY="344.0" prefHeight="29.0" prefWidth="831.0" progress="0.0" />
       <Pane fx:id="panebar" layoutX="73.0" layoutY="299.0" prefHeight="90.0" prefWidth="883.0" />
